requirements.txt
# core ui framework
streamlit

# image processing
Pillow

# data processing
pandas
